% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{cran_registry}
\alias{cran_registry}
\alias{cran_registry_with_status}
\alias{cran_registry_update_json}
\alias{update_crantogit_csv}
\title{CRAN tools}
\usage{
cran_registry()

cran_registry_with_status(full_reset = FALSE)

cran_registry_update_json()

update_crantogit_csv()
}
\arguments{
\item{full_reset}{do not keep data from current `crantogit.csv` in case
of errors.}
}
\description{
Scraping git repositories for CRAN packages
}
