% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{cran_registry}
\alias{cran_registry}
\alias{bioc_registry}
\alias{cran_registry_with_status}
\alias{cran_registry_update_json}
\title{CRAN tools}
\usage{
cran_registry()

bioc_registry()

cran_registry_with_status()

cran_registry_update_json()
}
\description{
Scraping git repositories for CRAN packages
}
